Added GTest infrastructure, have init_hash working
Using some mocked callbacks, I have the init_hash function generating a
valid hash table. Now I need to port the rest of the methods and add
more tests for correctness.

+++ api/common/types.h
@@ -32,8 +32,8 @@ typedef struct memory_management_functions {
  * Section: Method prototypes needed globally.
  ******************************************************************************/
 
-typedef int (*writeback_fn_t)(paddr_t, off_t, size_t, const char*);
-typedef int (*read_fn_t)(paddr_t, off_t, size_t, char*);
+typedef ssize_t (*writeback_fn_t)(paddr_t, off_t, size_t, const char*);
+typedef ssize_t (*read_fn_t)(paddr_t, off_t, size_t, char*);
 typedef ssize_t (*alloc_metadata_t)(size_t, paddr_t*);
 typedef ssize_t (*alloc_data_t)(size_t, paddr_t*);
